% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfunc.R
\name{newtonMethodR}
\alias{newtonMethodR}
\title{Using Newton's method to find the root of a function}
\usage{
newtonMethodR(f, df, x0, maxiter = 100, tol = 1e-06)
}
\arguments{
\item{f}{Function that need to solve roots}

\item{df}{The derivative of function f}

\item{x0}{Estimating the initial value of roots}

\item{maxiter}{Maximum number of iterations, default to 100}

\item{tol}{Error tolerance, used to determine the accuracy of the solution, defaults to 1e-6}
}
\value{
a random sample of size \code{n}
}
\description{
Implement a function optimized using Newton's method to solve the roots of nonlinear equations
}
\examples{
\dontrun{
f <- function(x) x^3 - x^2 - x - 1
df <- function(x) 3*x^2 - 2*x - 1
root <- newtonMethodC(f, df, x0 = 1.5)
print(root)
}
}
