\name{SA23204194-package}
\alias{SA23204194-package}
\alias{SA23204194}
\docType{package}
\title{
  Algorithm and example functions for K-means algorithm and NewtonMethod
}
\description{
   R and Rcpp versions of k-means and newtonmethod function are provided in this package (\code{kmeans_customR} and \code{newtonMethodR} versus \code{kmeans_customC} and \code{newtonMethodC})
}
\author{
Xu Tongzhou
Maintainer: Xu Tongzhou <xtz0716@mail.ustc.edu.cn>
}

\keyword{ package }

